# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News project API",
        "version": "1"
    },
    "paths": {
...
        "/login": {
            "get": {
                "tags": [
                    "Authenticated Session"
                ],
                "summary": "Display the login view",
                "description": "Displays the login view for the user to enter their credentials.",
                "operationId": "login",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Authenticated Session"
                ],
                "summary": "Authenticate the user",
                "description": "Authenticates the user and creates a session.",
                "operationId": "authenticate",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authenticated Session"
                ],
                "summary": "Destroy the authenticated session",
                "description": "Destroys the authenticated session and logs the user out.",
                "operationId": "logout",
                "responses": {
                    "302": {
                        "description": "Redirect"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/email/verification-notification": {
            "post": {
                "tags": [
                    "Email Verification"
                ],
                "summary": "Send a new email verification notification",
                "description": "Send a new email verification notification to the authenticated user.",
                "operationId": "7c01a354a5960445207bc5d2f17d547e",
                "responses": {
                    "302": {
                        "description": "Redirects to the previous page with a success message.",
                        "content": {
                            "text/html": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden error, user has already verified email.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User has already verified email."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/email/verify": {
            "get": {
                "tags": [
                    "Email Verification"
                ],
                "summary": "Display the email verification prompt.",
                "description": "Displays the email verification prompt if the user has not verified their email address yet.",
                "operationId": "getEmailVerificationPrompt",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmailVerificationPromptResponse"
                                }
                            }
                        }
                    },
                    "302": {
                        "description": "User has already verified their email",
                        "headers": {
                            "Location": {
                                "description": "Redirects to the intended page",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Verification"
                ],
                "summary": "Mark the authenticated user's email address as verified",
                "description": "Mark the authenticated user's email address as verified",
                "operationId": "1b8f267ccce552fda2edd55aab4e1711",
                "responses": {
                    "200": {
                        "description": "Email address has been verified",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email address has been verified"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "EmailVerificationPromptResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Email verification prompt displayed"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}